Suprise, suprise... more NEW_SOUND work (and a NEW_WEATHER test). Getting closer!
Fixed some minor bugs with actor sounds
Fixed a bug with the streams updating from the camera position, not actor pos
Added a variable point size for NEW_WEATHER rain when zoomed in - shall see if it meets Roja's approval
Added support for checking the available number of sources from OpenAL
Added support for streams fading in and out
Remodelled the arrays so sounds_list is the master list - it tracks the cookie - rather than a combination of it and
sources through different variables. This makes the code cleaner in the long run and allows for sounds to be added and
removed correctly while disabled.
Moved the destroy functions into the init section to group things a little better
Tidyed up the error messages in the init and config sections