Added missile types to the client: - added a new missile_type structure that contains properties of different missile types - added a shield_part structure to the actors code that contains one more integer refering to the type of equipped missiles - modified the code that loads the actors definitions from XML files to take in account the new shield_part structure - added a new file that contains missiles definitions: actor_defs/missile_defs.xml - added effects for special arrows: magic, fire, ice, explosive
Note: with this addition, all actors files (actor_defs/*_{fe}?male.xml)
have to be modified to add a <missile> tag in the shield part of the
quivers. This tag should contain the id of the missile type to use and
that has been defined in the actor_defs/missile_defs.xml file.