Don't allocate calculation stack separately
The calculation stack is currently stored as a pointer in its containing
struct, and is allocated separately. This might make sense if the stack needed
to be resizable, but since it's never reallocated anyway, we might as well
incorporate the array in the structure definition directly and get rid of an
unneeded allocation.