Don't allocate every token on the heap.
Currently the #calc allocates a separate token struc for every token it finds
in an expression, stores the pointers on a stack, and frees the tokens once
they are processed. This is unnecessarily slow and error-prone, simply allocate
the stack once as an array of tokens, and fill the array as needed.