Return early when widget is NULL in vscrollbar_drag()
vscrollbar_drag() checks if the pointer to the widget that it receives as an
argument is NULL before it determines the window it is part of, and continues
if it is (with the window set to NULL). However, on the next line it calls
vscrollbar_click() with the same widget pointer, which dereferences it. So,
return early if the pointer is NULL, it doesn't do anything in that case anyway.