Display achievements even if no texture is found
Currently when no texture can be found for an achievement, it will not be
displayed at all. It seems that the intent of the author was to continue
execution after an if-statement, but due to the continue statement used, the
code ignored the rest of the loop body and continued with the next
achievement. Fix, and draw a question mark when no texture is found, same
as when the achievement is not recognised.