Add support for kdialog
Add support for using KDE's kdialog in the game launcher. Things aren't perfect: as in the terminal dialog, I
can find no way to add column headers to the server selection menu. Furthermore, it seems to be impossible to
specify a window width, and the default width chosen for the server selection screen is a bit too low.

It seems to work however, and may be of value for those that have KDE installed, and no GNOME packages.