Remove cache size limit
While putting a limit on the size of the cache may be a good idea, the current size limit is never enforced, and
as far as I can see, it never has been. Since as far as I know, nobody has ever complained about it, just remove
it altogether and simplify the code somewhat.