Remove total_size from cache structure
The total_size member of a cache structure is supposed to hold the total amount of memory allocated for the cache
itself and the data it contains. However, it is not used anywhere, except to update the total_size of the system
meta cache (a cache of all caches). Furthermore, it is not even completely correct, since it fails to account for
the memory allocated for the cache_item_struct structures that hold the cached items.

Since it's not used in current code, simplify bookkeeping and just remove the field.