Fix size accounting of 3d objects
When loading 3d objects, the code currently tries to to adjust the size of a 3d object in the cache for the
allocated materials list and vertexbuffers. Problem is that it tries to do so before the item is even added to
the cache. As a result, only the size of the core e3d objects is accounted for in the cache.

Add the item to the cache before loading its details to fix this.