Fix crash on map change.
When changing maps, the current scene is stored as a texture, and the view is "frozen" by drawing only this
texture. To store the view, a texture is generated using glTexImage2D(). The last parameter to this function is
a pointer to the pixel data to be stored in the texture, but currently a pointer to the OpenGL texture handle is
passed instead. How in the world this has worked so far is a mystery to me, but it started crashing when I started
to compile in the actor definitions (thus changing the memory layout of the executable).

We don't actually need to copy any pixels from memory though, but from the current OpenGL read buffer. So pass NULL
as data pointer to prevent copying random bits of data.