Add option to add all TTF fonts from a directory to the font lists.
This needs more work. Adding all fonts on startup is slow as all fonts need te be read
and converted into a texture. Besides that, the current system of using font numbers
only works well with static lists. If a user now changes the font directory, the old
font numbers will refer to different fonts, or be invalid altogether.