Fix last character on a line not being drawn.
While the glyph would correctly be deemed to fit by reset_soft_breaks(), draw_messages will in some cases not draw it because it does not take the actual glyph width into account, but rather the maximum width it can be. This cannot easily be fixed, since draw_messages might also have to fit a cursor on the position.