Draw the separator line in the console in the same font size again
The font scale was not taken into consideration when drawing the lines of
carets that indicates the console can be scrolled down. This made the
characters in the line a different size than the characters in the chat text.
Revert this behaviour by taking the proper font scale into account.

I may change it back with a proper font scaling factor though. I think the
small carets look better.