Avoid creating context menu window before initialization.
Avoid the use of global static objects, as their initialization order is
undefined. In this case, the creation of an (in-game) window causes havoc
when it tries to access uninitialized data (like, say, fonts). Move the static
object to a class static function, where it will only be initialized on first
use.