Add font category to all widgets not just text fields.
This allows us to create e.g. labels or buttons with different font categories.
And also password fields. Furthermre, calculate the width of the string to be
displayed in a password field properly, and align the displayed string to the
right of the field if the width is too large. Finally, use the fact that password
fields can display partial input to reduce the width of the buddy name input
by not using the maximum character width for a font, but the average.