Attempt to fix vertical alignment
Some TTF fonts are rendered too high vertically. A possible cause may be that the
entire spacing between lines is added beneath the glyphs when generating the font
texture. Try to fix this by shifting the glyphs down by half the difference between
the font height and the line skip, so that half the empty space gets added to the
top. It seems to work for the Candara Bold font, but needs some testing to see if
it does not mess up other fonts.