Introduce unifying text drawing function.
Replace the multitude of text drawing function with a single unifying text
drawing function that can take a variable number of options. The implementations
of the previous functions have been replaced by inline calls to draw_text(),
unless only a single caller exists, in which case the call has been replaced by a
direct call to draw_text() and the function is removed.