Fix book text overflow.
When an image was placed after the current text page, and the image was taking
up an entire page or more, multiple pages need to be skipped. Only a single page
was skipped, and due to signed->unsigned conversion, the entire rest of the text
was placed on this page. Fixed now.