Fix write outside buffer
Off-by-one error in checking the length of the last_complete buffer in the tab
completion function can result in writing one byte past the buffer for long
enough inputs. Unlikely to trigger in normal usage, as a user would have to use
tab completion on a 48 byte or longer string without spaces. Fix by checking the
bound correctly.