Shift texture coordinates of bundled fonts up by one pixel.
At least for the default font, currently we include an empty pixel at the bottom
of every character. On the other hand, the ring at the top of the Å character is
not drawn completely. The result is that currently font 1 looks like it is shifted
up a bit. Note that this is not new behaviour, the current master branch does the
same. Fix by shifting the texture coordinates up by one pixel.