Fix keypress handler for integer spinbuttons.
Currently, after a keypress, the integer spinbuttons calculate the new value by
calculating 10 times the current value + the integer value of the newly inserted
character. This works, unless the user starts deleting characters. In that case,
the current value is not altered if the result would be lower than the minimum
value, and is no longer equal to the value depicted by the spinbuttons text. Fix,
by simply always converting the entire text to calculate the value.