Fix invalid memory access in encyclopedia.
On reading the first category, the height of the current page is checked, even
though no page has been added yet. This results in an access outside the pages
array. Fix by checking if the page number is valid before trying to update the
page height. Bug found by gcc's address sanitizer, which is also added as an
option in the Linux Makefile with this commit.