Fix length check on stats data.
The length check on the stats data sent by the server is out of date, the client
will happily access beyond the checked length. Fix the check to match what the
client actually uses. This will not be a problem in practice, as the server will
hopefully send the correct data, but let's make sure the check covers what we need.

Interestingly, it looks like the server actually sends 4 bytes more than what is
used by the client, wonder what these are for?