Keep two rows of empty pixels between glyphs in TTF textures.
Apparently on some systems, texture coordinates are off by can be off by a pixel.
This has been seen to result in parts of the character above being drawn when
that extends all the way to bottom. By keeping some empty space between rows
of characters we avoid this. Texture coordinates being off may still result in
part of the gyph being cut off for very low (or high) characters, but the effect
should be rather minimal, and for the vast majority of characters there should be
no cutoff at all.