Don't scale the notepad window when the notes font changes.
The interface only needs to change with changes to the UI font, so there is no need
to close the notes upon changing the notes font. Rewrapping the notes text is already
handled by the text widget, so there's nothing further to be done to handle note font
changes gracefully.