The TTF code changed the notepad popup handler to use the password widget.  This introduced a crash and a memeory leak.
To reproduce the crash beore this fix:
1. Open the notepad.
2. Click the "New Category" button.
3. Enter some text for the name but press cancel.
4. Click the "New Category" button again.
5. Note the previously entered text is still present.
6. Attempt to delete the text using the back delete key.
7. The client crashes.

The memory leak is with the initial allocation of the password widget
text buffer if the popup window was never opened.

This commit provides a new pword_clear() function that is called
instead if just setting the password text buffer to an empty string.
It fixes the memory leak by freeing the memeory even if the popup
window has not been opened.

@gvissers feel free to find a better fix.