Fix crash on startup.
When using the old chat interface, the client tries to wrap text early in the
startup process. This loads the font for the message, and as a result, tries
to load the texture for the font before the texture cache is initialized.
Wrapping the text at this point is unnecessary, since the messages will be
rewrapped when the console window is initialized, and new messages are wrapped in
update_console_win(). So remove the call to rewrap_message() to fix the crash.