Fix height of digital clock, and keep hours and minutes in place.
As the font size of the digital clock scales with the hud width, the height
of the time stamp should also be dynamic. So calculate the zoom factor for the
text beforehand, and use this factor to calculate the height correctly.

Furhtermore, when showing seconds using a variable width font, the width of
the time stamp can change every second. This cause the entire string to jump
left and right which is visually distracting. Fix the position of single
minutes character, so that the hours and minutes characters only move once
a minute to reduce the amount of visual noise..