Might as well center the minutes digit properly.
In the previous commit, the time string was shifted right by half the maximum
digit width in an attempt to center the middle digit. Since we have the digit
already printed in a string, we might as well shift by exactly half the actual
digit width, so that the center character is properly centered.