Fix numbers in scaled down inventory not showing up.
When the combined scale factor for text (window scale factor * text scale factor)
becomes smaller than 0.5, the shadow width is rounded down to 0 pixels. This in
itself is not wrong, but an error in the logic prevented the foreground text from
being drawn as well. Fix by only not drawing the shadow when the size becomes too
small.