Increase compatibility with "old" fonts.
Before the TTF merge, the font rendering would do an interesting dance, which for
the default fixed font and default font size would boil down to:
* In the horizontal direction, scale character widths by 11/12
* In the vertical direction, at default zoom the scaling cancels out, and the
  full 19 pixels are used[1], but when going to the next line the position is
  only advanced by 18 pixels.

The new font handling code tried to get away with a single scale factor for the
text, but as with any change, this broke some people's workflow (obligatory
reference: https://xkcd.com/1172/). So this patch introduces separate scaling
factors for the x and y directions, and a vertical advancement parameter that
determines the line spacing. This will make the characters in old fonts a bit
taller vertically without affecting the line height, making it look more like
the old rendering.

This patch should have no effect for TTF fonts.

[1]: 19 pixels, but the pixels selected included an empty pixel row at the bottom,
and possibly chopped off a non-empty pixel at the top. This was changed in the TTF
merge, and will not be reverted. Tailoring to people's workflow only goes so far.