Replace input fields in login window by password widgets.
This way, the fields can scroll to the right, and don't have to stretch to the
width needed to hold the widest possible username. This also adds cursor support
to the input fields.

To make this work, new functionality was added to the password widget:
* a new PWORD_FIELD_NO_BORDER flag to prevent drawing a border
* PWORD_FIELD_DRAW_CURSOR/PWORD_FIELD_NO_CURSOR flags to force/suppress drawing a cursor.
  Normally, a cursor is drawn when the mouse is over the password field. In the
  login window, we want to draw a cursor when the input field has focus, regardless
  of mouse cursor position.
* PWORD_FIELD_NO_KEYPRESS flags to prevent the field from handling text input.
  Keypresses are in first instance handled by the login window itself; only when
  the keypress is deemed acceptable, the PWORD_FIELD_NO_KEYPRESS flags is removed
  and the keypress is sent to the password widget that currently has focused.

Apart from cursor and scrolling support, the login window should work the same as before.