Scroll to selected option in multiselect.
When an option is selected in a multiselect widget, and the newly selected
option is not visible (presumably because the option is slected from code
instead of a click by the user), scroll to the selected option.

Implementing this revealed that th configuration window sets the value of
all multiselects every time the window is drawn. Apart from being a tad excessive,
this is actually harmful as the selects could now no longer be scrolled away
from the selected value to choose a new one. As far as I can see it was also
wholly unnecessary, so this code was removed.