Small font tweaks
* remove unused _block_height member in font
* Make unscaled height of TTF fonts 19 pixels instead of 21. The previous value
was based on the block height in the EL bundled font textures, but 2 of those
21 pixels are discarded. Note that the line height for TTF fonts is still the same
as the font height, not one pixel less as for the old fonts.
* Increase the ratio between small and normal font sizes a bit. Previously, the
ratio was based on the character width only, but the ratio between small and normal
characters scales differently in the x and y directions in the old font rendering.
The new value is a compromise between the ratios for the x and y directions. Or
for the cynics among us: now the small characters are both too wide and still not
high enough :)