Remove color from widgets constructors
Currently, all calls creating a widget explicitly pass the color of the widget
in their parameter list. But the vast majority of widgets (>90%) have the same
default color. Passing this same color for every constructor call is tedious,
error prone, and makes the parameter lists even longer than they need to be
(text_field_add_extended() has 17 parameters).

This patch creates widgets always in the default color. Widgets that need a
different color can call widget_set_color() afterwards. In some places the widget
is to be drawn in the current pen color. This is indicated by an invalid color
value. For these cases, the widget_unset_color() function is added.