Tidy up map window.
Currently the map window contents are stretched to occupy the full available
height, and 5/6 of the avilable width. This is ok if the game window's aspect
ratio is close to 3/2, but otherwise the map looks stretched or squashed,
depending on the video mode used. This not only affects the map itself, but
also any text drawn in the window.

This patch reworks the map window so that the correct aspect ratio is kept,
and centers it in the window, adding space if where necessary.