Change scaling of map marks
The map window is interpreted as having size 250x200, and so the font must be
scaled proportionally. This lead to having to set a small mapmark font size to
obtain the same font size in pixels as the normal UI font at size 1. Fix this
by incorporating the scale factor between screen coordinates and map coordinates
into the scaling factor, so that the map mark font size can be compared directly
to the UI font size.