Get rid of map window scaling.
Currently, screen coordinates in the map window are scaled down to a fixed
width of 250 and height 200. This wreaks havoc on the text, as font coordinates
are snapped to integer values, and the given width and height simply do not
provide enough resolution. Get rid of the scaling, and use pixel coordinates
explicitly.