Try to guess point size
Currently fonts are opened in a rather larger point size (40), and then scaled
down. This is not optimal if the scaling factor becomes large. In this commit
we try to find a point size, such that the font texture generated does not need to
be scaled for the default line height (times global UI scale factor). Also, the
text outline, which was fixed at two pixels, is now scaled with the global UI
scale.