Fix memory leak.
Fix small memory leak, caused by allocating the lines array in the selection
info of a text widget twice, and forgetting about the first allocation. As part
of the initialization of a text widget, its font category is set, and this calls
the font change handler. A recent change made the font change handler reallocate
the lines array, which was needed because the number of lines can change
depending on the font and size. Further on in the initialization of the text
widget, the array was explicitly allocated again, and the first allocation by the
font change handler was discarded. Fix by no longer allocating in the
initialization handler.