Remove pointless null termination.
In put_mark_on_position(), the byte at index strlen(marks[max_mark].text) is
set to zero to terminate the string. But if this byte was nt already zero, the
call to strlen() would return an incorrect, undefined, result. So this
assignment is incorrect. Since both the safe_strncpy() and the rtrim_string()
calls before it correctly terminate their results, it is unnecessary as well.