Fix trade window text alignment when going to storage. The changes for the new font rendering changed the alignment of quantities and now overwrote the "sto" text, used when items are marked as going to storage. Now,  the text is simply a single "S" to use less space and right aligned at the top or bottom of the grid slot opposite to the quantity text.
Fix existing calls not ending the parameter list with TDO_END.
The "S" is displayed using a new draw_text() function and I copied this
form the existing draw_text() calls in the same function.  However at some
scaling value the text did not show.  Much time debugging later, I found this
was due to the function call not terminating the parameter list with the,
TDO_END value.  The existing functions did not do this and so I've now fixed
these too.  A quick scan of other usage of draw_text() show they are all fine
and I was just unlucky.