Issue #86: Fix inconsistencies with window action modes.
Previously:
The quick-bar window, inventory window and game root window had separate
action  modes.  The ground bag shared the inventory window mode, the
trade window and manufacture window shared the game root window.
Casting a spell would reset the game root window mode to ACTION_WALK,
whether or not a change to ACTION_WAND was needed or the spell failed.
All three mode settings were changed when a action icon was clicked in
the icon window or one of the action keys pressed.  For some actions the
selected modes were not valid for the quick-bar or inventory windows, so
they reset the window modes to the default.

The changes:
  The quick-bar window, inventory window and game root window still
have separate modes but there are options to prevent an action icon or
keypress from changing the quick-bar or inventory window modes; these
options make the windows totally independent.  Even if the option is not
set, only usable modes will now change the quick-bar or inventory window
settings.
  Casting a spell only changes the game window action mode if a target
is needed for the spell (requiring the wand cursor).  Other spells, or
a failed spell do not change the action mode.
  The ground bag in addition to the manufacture and trade windows, now
share the game root window action mode.  They now all behave the same in
that a right click will show the item description.  If the EYE action is
active, the EYE cursor is displayed.  Right-clicking will cancel any
dragged items, or cancel any active eye action mode.  This is a minor
change to the ground bag window and a bug fix for the trade window (it
did not previously show the EYE cursor).  The manufacturing window
behaviours as it did previously. The only time the game root window
action is changed by these windows is when they cancel an active EYE
mode.

Overall, this should see more consistent behaviour of action modes, and
unnecessary changes to the game root window are removed.  The quick-bar
and inventory windows can be total independent by choice.