Remove my_strncp()
It provides exactly the same functionality as safe_strncpy(), but in an open coded loop.
Furthermore, it contains a small mistake: because len is decremented before it is checked,
and is unsigned, calling my_strncp() with len==0 will cause it to try and copy 2^64-1 bytes,
guaranteeing a segfault.

Replace all calls to my_strncp() with safe_strncpy().