Remove my_strcompare()
The function my_strcompare() acts somewhat like the standard libc's strcasecmp except
that it inverts the logic, and does not return an ordering, only 1 if the strings are
equal and 0 if not. It really should have been called my_strequal() or similar. Also,
it is implemented rather inefficiently, first calculating and comparing the lengths of
both strings before comparing the contents, thus needing to iterate through the strings
twice if the lengths are equal.

Replace my_strcompare() with !strcasecmp(), except in new_actors.c. I believe the use
of !my_strcompare() in add_enhanced_actor_from_server() was wrong, and the intent was
to check for equality, not difference, of both actor names. Unfortunately this is rather
hard to check as it involves two actors of type COMPUTER_CONTROLLED_HUMAN, which, if I
am not mistaken, only occurs when a player disconnects but is still in combat.