First attempt at fixing overtext
Try to fix overtext, aka speech bubbles. Overhead speech and bubbles are now drawn
in orthographic projection instead of in perspective view, making it easier to reason
about coordinates and scaling. It also correctly uses the chat font now.

This is not completely finished. The location of the bubbles probably needs some
fine tuning, and probably the distance scaling as well. But the basic feature works.