Fix map editor build
Previous commits that included chat.h from actors.h broke the map editor build. As the map editor
does not even use actors, this should not be necessary. The only place actors.h was included in the
map editor was through eye_candy_wrapper.{h, cpp}. Update these files to only use actors when not
compiling for the map editor. Wile we're at it, make a few objects and definitions local that do not
need to be exported, and remove a few unused functions.