Fix width calculation fof text in overhead bubbles
Previously, the width of the text in the bubbles was calculated at a certain zoom level, and then
scaled up or down based on the distance of the actor to the camera. This does not work very well,
as text characters are aligned on integer pixel coordinates, so that the width of each character
should be rounded before adding to the text width. Fix by computing the bubble width in real time
from the actual string with the correct scaling factor.