Fix missing 2D objects on snow map
2D object can have an alpha test parameter, where fragments with alpha less than the parameter
specified in the object are discarded. With the current multitexturing setup for cloud shadows,
alpha combining of the various textures could lead to lower alpha values than the object texture
itself holds. This is especially noticable on snow objects, where the alpha threshold is very
high (0.9), and so they disappeared when cloud shadows were enabled. Fix by explicitly setting
the texture combining parameters to always return the alpha value of the object texture.