Fix compilation with PAWN
Make the client compile again when PAWN is enabled. It ...works...: the tree near
beam turns every 2 seconds, check it out! :P) Presumably Derzelas will also turn towards
you when you enter his room, but I have not tested this.

I am really not sure what to do with this feature, however. Radu wanted it at some point
to send scripts from the server to be executed to the client, and it could perhaps in theory
work, but
1) There are perhaps three people in the world that even know Pawn the scripting language exists,
2) It is not all that difficult to learn, but it does have its limitations, Nicer scripting languages exist.
3) I don't think server support is going anywhere, and
4) Noone has ever used it, or asked for a scripting environment in the game AFAIK

Perhaps it would be better to just scrap it altogether. I will miss my turning tree, though.