Replace postfix increment on iterators
Using a postfix increment or decrement operator requires the old value to be stored before changing
the value, so that it can be returned later. This is usually not a problem for simple integers or
pointers, but can possibly have a performance impact for more complex iterators.

Replace loops with postfix increments by range-based for loops where possible, and use prefix operators
where it is not.