Harden string settings in elconfig
Two changes in string handling in elconfig, to try and prevent overflows:
1) Most string options in elconfig were using hard-coded buffer lengths. Since they are all declared
"extern char str[SOME_SIZE]", the buffer size is known and we can simply use sizeof(str) instead.
2) the change_string() function takes a length option (to which the stored buffer size is passed),
but will happily copy that many characters and put a terminator after that. This is contrast to other
string handling functions in the client, that include the terminator in the character count. Fixed
by simply calling safe_strncpy() instead of hand-coding a copy loop.