Fix buffer overrun in password field.
Clicking in the small space befire the first character could cause the cursor position to become less
than the position of the first character drawn (-1 if the first character was drawn). This lead to
a negative buffer size being passed to get_buf_width_zoom(), which was interpreted as a very large
unsigned value. Hence, the buffer was overrun when checking the string width. Fix by returning clipping
the allowed cursor position to the index of the first drawn character.