Fix length check on incoming storage updates
The incoming length for a storage update is at least two bytes. Furthermore, if the update is
an actual update instead of a complete storage list, it should contain at least a single full
packet. Fix the length check in multiplayer.c, and add an extra check in storage.c for updates
so that the client does not read past the buffer on outdated or malicious servers.