Fix input field in map editor save dialog
The callback handler for keypresses for GTK dialogs should return a boolean that indicates whether
the event should be processed further. As the function was declared void, the return value was undefined
and could prevent the input character from being added to the input field. Fix the return type, and
return FALSE to indicate the event should be processed by the dialog handler.