Fix grid height control in map editor.
Issue 98: after opening and closing the ground tile window in the map editor, the grid can no longer be
controlled by the +/- buttons. This is due to the fact that the variable that keeps track of whether the
ground tile window is apparently not reset properly when the window is closed. Fixed by removing the
variable altogether, and using get_show_window() to check if the tiles window is open instead.