Change object name for deleted objects in map editor.
Issue #104: When an object is deleted from the map, its "blended" field is set to 20, but the rest of the structure
is kept intact. Such objects are not meant to be used, but there is a small possibility they might be. Change the
object file name for deleted objects to that of the "bad" object which is never used for anything to make doubly sure
they are never used.