Support loading TTF fonts from the game data directory.
TTF font paths are now first interpret as being relative to the game data directory, if that fails as a
path relative to the user specified font directory, and lastly as an absolute path. Furthermore, on
TTF initialization, the datadir/fonts directory is checked for TTF files. This allows us to bundle
TrueType fonts with the game.