Improved validation of password changes.
The #change_pass text was previously sent unchecked to the server. Where
a space was included in the new password, the server just used the first word
but the password manager silently rejected the change as it was invalid.  Now, the
#change_pass text is validated before sending to server, if it is not valid then
information is shown and it will not be sent to the server. This validation
means the password manager will never have to reject a #change_pass string
that the server has used.  But it will log an error if it does.

Also now show an error message on the login screen if login attempt causes a
disconnect from server.  This was failing silently before and so was confusing
to the user.