Do not strip internal spaces in config values
The code reading el.ini strips all spaces from the input values that are not quoted, also
spaces inside those values. This is a problem for the font multiselects which store the
font file name in parentheses after the numerical index, because those file names may contain
spaces. Fix the code to only remove spaces from the front and back of the string, but not
inside it.