Fix config value string offset
The previous commit (fa5a12b) copied a config value to a temporary buffer, but unlike the code prior to that,
neglected to update the (descriptively named) "ptr" variable to said buffer. Since
a) before the copy, "ptr" is already pointing to the correct offset in the (source) variable string, and
b) in the other part of the branch, we are also modifying the source string,
forego the string copy altogether, and simply modify the source to put a string terminator '\0' at the
end of the stripped variable value.