Fix display of outlined characters
When generating the texture for a TTF font with outline, the glyphs are pushed down in the textureby the size
of the outline. For some low-descending characters, the bottom of the glyph can then be overwritten by the
next row in the texture (and besides that, the bottom of the glyph was not properly stored in the texture
coordinates anyway). Reserve space for the outline in the texture, adjust texture coordinates to take the outline
into account, and similarly for the character drawing function.