Adjust glyph positions only if offset is negative
TTF fonts store the minimum x coordinate of the glyphs in the font. If this coordinate is negative, the
glyph extends to the left of the pen position, and we should adjust for that when drawing. However, we
should not adjust for positive offsets, as these indicate spacing around a glyph which should be included.