Fix wiggling of banner text
In commit f026063 a horizontal offset was added for banner texts. In the process, a temporary x coordinate
was calculated of type int, losing the fractional part of coordinate. This caused the text to jump from one
integer value to the next, instead of smoothly following the actor's motion. Fix by declaring the offending
variable as float.