Check for too small font sizes.
When the line height is too small, TTF fonts can be loaded with very small point sizes which in turn can lead to
"empty" glyphs with zero width. This will trigger an error in SDL_ttf, which will mark the font as failed and
reset the font for the category it is in. This also affects all other strings which are drawn in that category,
even when these are drawn at normal sizes, which is obviously undesirable. It is hard to distinguish an error
resulting from a too small font size from "real" rendering issues, so we introduce a minimum point size for
TTF fonts, which should be safe from rendering issues. If the font size is too small, the glyph textures
will be downscaled, but at least the font will not fail to load.