Fixes to my_isupper
Two fixes for my_isupper. The first is checking the len parameter after testing the first 3 characters for a
string terminator. Passing in a short (< 3 byte) buffer without a '\0' byte could lead to an invalid memory
access. The second is a slight change in the behaviour of the function. In the old situation, when a digit
is found and the number of alphabetic characters *at that point* is less than half the length of the string,
0 (not shouting) is returned. Thus the filter is easily circumvented by starting the text with a digit
("0ALL CAPS!!1ELEVENTY ONE!"). Now the number of alphabetic characters is checked after the whole string has
been processed, so the total number of upper case alphabetic characters must be less than half of the string
length for the string to not be filtered.