Fixes for all caps filter
When filtering all caps text, the entire text line, including any names or "PM from" string, was lower cased
when the text behind the name was in all caps (or whatever test my_isupper() performs on the string). Lower
casing the name when it is not even tested is strange, and is inconsistent when in some chat lines the name is
in lower case where in others it is not. So only lower case the tested string if it is found to be shouting.
Furthermore, checks on the string index were performed after accessing the string at said index, which could
lead to an invalid memory access. This is fixed now.