(Nearly) get rid of get_string_occurrence()
The functionality of get_string_occurrence() is more or less equal to safe_strcasestr(), but it is open-coded.
Most users can be moved over to use safe_strcasestr() without trouble. The only exception is the braindead old
2D object loading code, which searches the entire content of the 2d0 file every time for every possible value
it may contain. This was fixed with FASTER_MAP_LOAD, but not for the map editor. Move the map editor over to use
the parse_2d0 function from the new 2d object loading code, and relegate get_string_occurrence() to !FASTER_MAP_LOAD,
where it can hopefully be deleted at some point.

This also fixes a bug in get_tile_sound(), where the string length of the needle was passed where the length
of the haystack was expected.