Don't add 1 to texture coordinates of 2D objects
In 451b873d the vertical texture coordinates in 2D objects were flipped, due to NEW_TEXTURES apparently reversing
the definition. In the process, the v coordinates, which were defined as 1-<some value> were changed to
1+<some value>, resulting in v coordinates greater than 1. This does not lead to problems, as the textures are
repeated when coordinates roll over 1, but it is also unnecessary. Remove the addition of 1 for a very minor
code simplification.