Make my_socket static
Every call to my_tcp_send() takes a socket parameter, and every one of them passes the same socket. Since it
seems very unlikely that the client will need to communicate with multiple game servers at the same time (and
besides, that would require a fair amount of rewrite anyway), remove this parameter from the function, and make
the socket local to multiplayer.c. This also affects my_tcp_flush(), from which the socket parameter can be
removed for the same reason, and load_empty_map(), which closes the connection on error. Add a new helper function
my_tcp_forced_quit() for the latter case.