Implement our own network code
If we want to encrypt the traffic with the server, we need to be able to work with the socket. SDL_net provides
a nice interface, but in the process abstracts away most of the details we need to be able to use SSL. So implement
our own socket type. Furthermore, split off the networking part of the message handling from multiplayer.c,
into a separate file, and use the new TCPSocket.

In time we may get rid of other uses of SDL_net in the code too. This is early code.